+++ src/TC/src/FfmpegDecoder.cpp
@@ -1,5 +1,7 @@
 /*
  * Copyright 2020 NVIDIA Corporation
+ * Copyright 2023 VisionLabs LLC
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -100,7 +102,11 @@ struct FfmpegDecodeFrame_Impl {
       cerr << "Could not find video stream in the input, aborting" << endl;
     }
 
-    auto p_codec = avcodec_find_decoder(video_stream->codecpar->codec_id);
+    auto entry = av_dict_get(pOptions, "c:v", NULL, 0);
+    auto p_codec = entry
+                       ? avcodec_find_decoder_by_name(entry->value)
+                       : avcodec_find_decoder(video_stream->codecpar->codec_id);
+
     if (!p_codec) {
       stringstream ss;
       ss << "Failed to find codec by id" << endl;
@@ -196,32 +202,6 @@ struct FfmpegDecodeFrame_Impl {
     return true;
   }
 
-  bool SaveGRAY12LE(AVFrame* pframe) {
-    size_t size = frame->width * frame->height * 2;
-
-    if (!dec_frame) {
-      dec_frame = Buffer::MakeOwnMem(size);
-    } else if (size != dec_frame->GetRawMemSize()) {
-      delete dec_frame;
-      dec_frame = Buffer::MakeOwnMem(size);
-    }
-
-    auto plane = 0U;
-    auto* dst = dec_frame->GetDataAs<uint8_t>();
-
-    auto* src = frame->data[plane];
-    auto width = frame->width;
-    auto height = frame->height;
-
-    for (int i = 0; i < height; i++) {
-      memcpy(dst, src, 2*width);
-      dst += 2*width;
-      src += frame->linesize[plane];
-    }
-
-    return true;
-  }
-
   bool SaveYUV444(AVFrame* pframe)
   {
     size_t size = frame->width * frame->height * 3;
@@ -295,8 +275,6 @@ struct FfmpegDecodeFrame_Impl {
       return SaveYUV422(frame);
     case AV_PIX_FMT_YUV444P:
       return SaveYUV444(frame);
-    case AV_PIX_FMT_GRAY12LE:
-      return SaveGRAY12LE(frame);
     default:
       cerr << __FUNCTION__ << ": unsupported pixel format: " << frame->format
            << endl;
@@ -419,9 +397,6 @@ void FfmpegDecodeFrame::GetParams(MuxingParams& params)
     break;
   case AV_PIX_FMT_YUV420P12:
     params.videoContext.format = P12;
-    break;
-  case AV_PIX_FMT_GRAY12LE:
-    params.videoContext.format = GRAY12;
     break;
   default:
     stringstream ss;
